requirements.txt
streamlit==1.42.2
langchain==0.3.19
langchain-openai==0.3.7
langgraph==0.3.2
openai==1.65.2
python-dotenv==1.0.1
wikipedia==1.4.0
beautifulsoup4==4.13.3
requests==2.32.3
faiss-cpu==1.10.0
tiktoken==0.9.0
langchain-community==0.3.18
sseclient==0.0.27
fastapi==0.115.11
uvicorn==0.34.0
matplotlib==3.10.1
langfuse==2.59.7
networkx==3.4.2
duckduckgo_search==7.5.4

# 추가된 의존성 (척추 유합술 상담 시스템용)
PyPDF2==3.0.1
pdfplumber==0.11.4
sqlalchemy==2.0.36
sqlite3  # Python 내장 모듈
pydantic==2.10.4
asyncio  # Python 내장 모듈
logging  # Python 내장 모듈

# 선택적 의존성 (성능 향상용)
chromadb==0.5.23
sentence-transformers==3.3.1
